Implement messaging improvements and new features
The commit introduces several changes to the BetterBanSystem codebase. These include improvements to the messaging system, addition of a new "automod" feature in the configuration file, as well as various related enhancements and refactoring efforts. It also includes several bug fixes and tackle issues to enhance overall application stability and performance.

Signed-off-by: Simon W. <58950483+SimonPlays15@users.noreply.github.com>