Refactor and extend plugin features
Majorly reworked and improved several areas of the BetterBanSystem plugin including the addition of new AutoMod functionality. A README.md was updated to explain the configuration of new features and to better explain previously existing settings. Additions include the ability to check for spam messages, duplicate messages, excessive caps lock usage, message links, and message profanity. The mute functionality was improved, logging was expanded upon and the dispatching of commands was improved across multiple files.

Signed-off-by: Simon W. <58950483+SimonPlays15@users.noreply.github.com>