Add update checker to BetterBanSystem
An `Updater` class was added, which checks for new versions of BetterBanSystem plugin from an external URL. This is triggered on system startup, and if a new version is detected, an informational message is logged after a delay of 4 seconds. The code responsible for the check has been encapsulated in a CompletableFuture for non-blocking execution.

Signed-off-by: Simon W. <58950483+SimonPlays15@users.noreply.github.com>