Add web service backend with security configuration
This update encompasses the creation of 3 new classes that set up a web service backend for the BetterBanSystem. A WebServiceApplication class was set up as the main entry point to the application, while WebSecurityConfig is responsible for the configuration of web security. A controller, WebServiceController, was also added to handle web requests. Additionally, added dependencies for spring-boot-starter-web and spring-boot-starter-security in pom.xml.

Signed-off-by: Simon W. <58950483+SimonPlays15@users.noreply.github.com>