Implement JWT Authentication & new User entities
The update includes the creation of JWT authentication and the introduction of new User entities. The "HomepageHandler" file was deleted as it is no longer needed with the new changes. Specific Vue.js components and views were added for user navigational purposes. Changes were also made to the "login" method and store constants have been updated.

Signed-off-by: Simon W. <58950483+SimonPlays15@users.noreply.github.com>